+++ axelor-business-project/src/main/java/com/axelor/apps/businessproject/service/batch/BatchInvoicingProjectService.java
@@ -55,13 +55,13 @@ public class BatchInvoicingProjectService extends AbstractBatch {
       TraceBackService.trace(e);
     }
 
-    List<Map<String, Object>> generatedInvoicingProjectList = new ArrayList<Map<String, Object>>();
+    List<Object> generatedInvoicingProjectList = new ArrayList<Object>();
 
     List<Project> projectList =
         projectRepo
             .all()
             .filter(
-                "self.isInvoiceable = ?1 AND self.statusSelect NOT IN (?2)",
+                "self.toInvoice = ?1 AND self.statusSelect NOT IN (?2)",
                 true,
                 ProjectRepository.STATE_CANCELED,
                 ProjectRepository.STATE_FINISHED)