+++ axelor-stock/src/main/java/com/axelor/apps/stock/service/StockMoveLineServiceImpl.java
@@ -452,9 +452,6 @@ public class StockMoveLineServiceImpl implements StockMoveLineService {
           qty = stockMoveLine.getQty();
         }
 
-        if (toStockLocation.getTypeSelect() != StockLocationRepository.TYPE_VIRTUAL) {
-          this.updateAveragePriceLocationLine(toStockLocation, stockMoveLine, fromStatus, toStatus);
-        }
         this.updateLocations(
             stockMoveLine,
             fromStockLocation,
@@ -465,6 +462,9 @@ public class StockMoveLineServiceImpl implements StockMoveLineService {
             toStatus,
             lastFutureStockMoveDate,
             stockMoveLine.getTrackingNumber());
+        if (toStockLocation.getTypeSelect() != StockLocationRepository.TYPE_VIRTUAL) {
+          this.updateAveragePriceLocationLine(toStockLocation, stockMoveLine, fromStatus, toStatus);
+        }
         weightedAveragePriceService.computeAvgPriceForProduct(stockMoveLine.getProduct());
       }
     }
@@ -541,7 +541,7 @@ public class StockMoveLineServiceImpl implements StockMoveLineService {
     } else {
       newAvgPrice = oldAvgPrice;
     }
-    stockLocationLine.setAvgPrice(newAvgPrice);
+    stockLocationLineService.updateWap(stockLocationLine, newAvgPrice, stockMoveLine);
   }
 
   @Override
@@ -1188,9 +1188,12 @@ public class StockMoveLineServiceImpl implements StockMoveLineService {
 
   public void fillRealizeWapPrice(StockMoveLine stockMoveLine) {
     StockLocation stockLocation = stockMoveLine.getStockMove().getFromStockLocation();
-    StockLocationLine stockLocationLine =
-        stockLocationLineService.getStockLocationLine(stockLocation, stockMoveLine.getProduct());
+    Optional<StockLocationLine> stockLocationLineOpt =
+        Optional.ofNullable(
+            stockLocationLineService.getStockLocationLine(
+                stockLocation, stockMoveLine.getProduct()));
 
-    stockMoveLine.setWapPrice(stockLocationLine.getAvgPrice());
+    stockLocationLineOpt.ifPresent(
+        stockLocationLine -> stockMoveLine.setWapPrice(stockLocationLine.getAvgPrice()));
   }
 }