+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/SaleOrderInvoiceServiceImpl.java
@@ -46,8 +46,11 @@ import com.axelor.apps.sale.db.repo.SaleOrderRepository;
 import com.axelor.apps.sale.service.app.AppSaleService;
 import com.axelor.apps.sale.service.saleorder.SaleOrderComputeService;
 import com.axelor.apps.sale.service.saleorder.SaleOrderLineService;
+import com.axelor.apps.sale.service.saleorder.SaleOrderWorkflowServiceImpl;
 import com.axelor.apps.stock.db.StockMove;
 import com.axelor.apps.stock.db.repo.StockMoveRepository;
+import com.axelor.apps.supplychain.db.Timetable;
+import com.axelor.apps.supplychain.db.repo.TimetableRepository;
 import com.axelor.apps.supplychain.exception.IExceptionMessage;
 import com.axelor.apps.supplychain.service.app.AppSupplychainService;
 import com.axelor.apps.supplychain.service.invoice.generator.InvoiceGeneratorSupplyChain;
@@ -90,6 +93,8 @@ public class SaleOrderInvoiceServiceImpl implements SaleOrderInvoiceService {
 
   protected StockMoveRepository stockMoveRepository;
 
+  protected SaleOrderWorkflowServiceImpl saleOrderWorkflowServiceImpl;
+
   @Inject
   public SaleOrderInvoiceServiceImpl(
       AppSupplychainService appSupplychainService,
@@ -97,15 +102,16 @@ public class SaleOrderInvoiceServiceImpl implements SaleOrderInvoiceService {
       InvoiceRepository invoiceRepo,
       InvoiceService invoiceService,
       SaleOrderLineService saleOrderLineService,
-      StockMoveRepository stockMoveRepository) {
+      StockMoveRepository stockMoveRepository,
+      SaleOrderWorkflowServiceImpl saleOrderWorkflowServiceImpl) {
 
     this.appSupplychainService = appSupplychainService;
-
     this.saleOrderRepo = saleOrderRepo;
     this.invoiceRepo = invoiceRepo;
     this.invoiceService = invoiceService;
     this.stockMoveRepository = stockMoveRepository;
     this.saleOrderLineService = saleOrderLineService;
+    this.saleOrderWorkflowServiceImpl = saleOrderWorkflowServiceImpl;
   }
 
   @Override
@@ -115,7 +121,8 @@ public class SaleOrderInvoiceServiceImpl implements SaleOrderInvoiceService {
       int operationSelect,
       BigDecimal amount,
       boolean isPercent,
-      Map<Long, BigDecimal> qtyToInvoiceMap)
+      Map<Long, BigDecimal> qtyToInvoiceMap,
+      List<Long> timetableIdList)
       throws AxelorException {
 
     Invoice invoice;
@@ -130,6 +137,34 @@ public class SaleOrderInvoiceServiceImpl implements SaleOrderInvoiceService {
 
       case SaleOrderRepository.INVOICE_ADVANCE_PAYMENT:
         invoice = generateAdvancePayment(saleOrder, amount, isPercent);
+        break;
+
+      case SaleOrderRepository.INVOICE_TIMETABLES:
+        BigDecimal percentSum = BigDecimal.ZERO;
+        TimetableRepository timetableRepo = Beans.get(TimetableRepository.class);
+        List<Timetable> timetableList = new ArrayList<>();
+        if (timetableIdList == null || timetableIdList.isEmpty()) {
+          throw new AxelorException(
+              saleOrder,
+              TraceBackRepository.CATEGORY_INCONSISTENCY,
+              I18n.get(IExceptionMessage.SO_INVOICE_NO_TIMETABLES_SELECTED));
+        }
+        for (Long timetableId : timetableIdList) {
+          Timetable timetable = timetableRepo.find(timetableId);
+          timetableList.add(timetable);
+          percentSum = percentSum.add(timetable.getPercentage());
+        }
+        invoice =
+            generateInvoiceFromLines(
+                saleOrder, this.generateQtyToInvoiceMap(saleOrder, percentSum), true);
+
+        if (!timetableList.isEmpty()) {
+          for (Timetable timetable : timetableList) {
+            timetable.setInvoice(invoice);
+            timetableRepo.save(timetable);
+          }
+        }
+
         break;
 
       default:
@@ -149,6 +184,38 @@ public class SaleOrderInvoiceServiceImpl implements SaleOrderInvoiceService {
     invoice = invoiceRepo.save(invoice);
 
     return invoice;
+  }
+
+  private Map<Long, BigDecimal> generateQtyToInvoiceMap(
+      SaleOrder saleOrder, BigDecimal percentage) {
+    Map<Long, BigDecimal> map = new HashMap<>();
+
+    for (SaleOrderLine soLine : saleOrder.getSaleOrderLineList()) {
+      this.addSubLineQty(map, percentage, soLine.getId());
+      map.put(soLine.getId(), percentage);
+    }
+    return map;
+  }
+
+  @Override
+  public void addSubLineQty(
+      Map<Long, BigDecimal> qtyToInvoiceMap, BigDecimal qtyToInvoiceItem, Long soLineId) {
+
+    SaleOrderLine saleOrderLine = Beans.get(SaleOrderLineRepository.class).find(soLineId);
+
+    if (saleOrderLine.getTypeSelect() == SaleOrderLineRepository.TYPE_PACK) {
+      for (SaleOrderLine subLine : saleOrderLine.getSubLineList()) {
+        BigDecimal qty = BigDecimal.ZERO;
+        if (saleOrderLine.getQty().compareTo(BigDecimal.ZERO) != 0) {
+          qty =
+              qtyToInvoiceItem
+                  .multiply(subLine.getQty())
+                  .divide(saleOrderLine.getQty(), 2, RoundingMode.HALF_EVEN);
+        }
+        qty = qty.setScale(2, RoundingMode.HALF_EVEN);
+        qtyToInvoiceMap.put(subLine.getId(), qty);
+      }
+    }
   }
 
   @Override
@@ -234,9 +301,13 @@ public class SaleOrderInvoiceServiceImpl implements SaleOrderInvoiceService {
 
     Invoice invoice = invoiceGenerator.generate();
 
-    invoiceGenerator.populate(
-        invoice,
-        this.createInvoiceLinesFromTax(invoice, taxLineList, invoicingProduct, percentToInvoice));
+    List<InvoiceLine> invoiceLinesList =
+        (taxLineList != null && !taxLineList.isEmpty())
+            ? this.createInvoiceLinesFromTax(
+                invoice, taxLineList, invoicingProduct, percentToInvoice)
+            : this.createInvoiceLinesFromSO(invoice, saleOrder, invoicingProduct, percentToInvoice);
+
+    invoiceGenerator.populate(invoice, invoiceLinesList);
 
     invoice.setAddressStr(saleOrder.getMainInvoicingAddressStr());
 
@@ -319,6 +390,54 @@ public class SaleOrderInvoiceServiceImpl implements SaleOrderInvoiceService {
       }
     }
     return createdInvoiceLineList;
+  }
+
+  protected List<InvoiceLine> createInvoiceLinesFromSO(
+      Invoice invoice, SaleOrder saleOrder, Product invoicingProduct, BigDecimal percentToInvoice)
+      throws AxelorException {
+
+    List<InvoiceLine> invoiceLineList = new ArrayList<>();
+    BigDecimal lineAmountToInvoice =
+        percentToInvoice
+            .multiply(saleOrder.getInTaxTotal())
+            .divide(new BigDecimal("100"), 4, BigDecimal.ROUND_HALF_EVEN);
+
+    InvoiceLineGenerator invoiceLineGenerator =
+        new InvoiceLineGenerator(
+            invoice,
+            invoicingProduct,
+            invoicingProduct.getName(),
+            lineAmountToInvoice,
+            lineAmountToInvoice,
+            lineAmountToInvoice,
+            invoicingProduct.getDescription(),
+            BigDecimal.ONE,
+            invoicingProduct.getUnit(),
+            null,
+            InvoiceLineGenerator.DEFAULT_SEQUENCE,
+            BigDecimal.ZERO,
+            PriceListLineRepository.AMOUNT_TYPE_NONE,
+            lineAmountToInvoice,
+            null,
+            false,
+            false,
+            0) {
+          @Override
+          public List<InvoiceLine> creates() throws AxelorException {
+
+            InvoiceLine invoiceLine = this.createInvoiceLine();
+
+            List<InvoiceLine> invoiceLines = new ArrayList<>();
+            invoiceLines.add(invoiceLine);
+
+            return invoiceLines;
+          }
+        };
+
+    List<InvoiceLine> invoiceOneLineList = invoiceLineGenerator.creates();
+    invoiceLineList.addAll(invoiceOneLineList);
+
+    return invoiceLineList;
   }
 
   @Override
@@ -611,6 +730,11 @@ public class SaleOrderInvoiceServiceImpl implements SaleOrderInvoiceService {
           saleOrder.getSaleOrderSeq());
     }
     saleOrder.setAmountInvoiced(amountInvoiced);
+
+    if (appSupplychainService.getAppSupplychain().getCompleteSaleOrderOnInvoicing()
+        && amountInvoiced.compareTo(saleOrder.getExTaxTotal()) == 0) {
+      saleOrderWorkflowServiceImpl.completeSaleOrder(saleOrder);
+    }
   }
 
   @Override